dapeng-2.0.2

+++ dapeng-api-doc/src/main/java/com/github/dapeng/doc/TestController.java
@@ -91,11 +91,12 @@ public class TestController {
         }
 
         if (parameters.contains("operatorId")) {
-            invocationCtx.operatorId(Long.valueOf(req.getParameter("operatorId")));
+            invocationCtx.operatorId(Integer.valueOf(req.getParameter("operatorId")));
         }
 
         if (parameters.contains("userId")) {
             invocationCtx.userId(Long.valueOf(req.getParameter("userId")));
         }
+
     }
 }
